backend/requirements.txt
Django==5.0.2
pandas==2.2.1
openpyxl==3.1.2
djangorestframework==3.14.0
markdown==3.5.2
django-filter==23.5
coverage==7.4.3
django-cors-headers==4.3.1
celery==5.3.6
redis==5.0.2